hireing status and resume qrcode
